+++ kedro_airflow_k8s/operators/node_pod.py
@@ -39,6 +39,7 @@ class NodePodOperator(KubernetesPodOperator):
         node_selector_labels: Optional[Dict[str, str]] = None,
         labels: Optional[Dict[str, str]] = None,
         tolerations: Optional[List[Dict[str, str]]] = None,
+        annotations: Optional[Dict[str, str]] = None,
         source: str = "/home/kedro/data",
     ):
         """
@@ -60,7 +61,10 @@ class NodePodOperator(KubernetesPodOperator):
         :param requests_memory: k8s requests memory value
         :param limits_cpu: k8s limits cpu value
         :param limits_memory: k8s limits memory value
-        :param node_selector_labels: dictionary of labels to be put into pod node selector
+        :param node_selector_labels: dictionary of node selector label to be put into pod node selector
+        :param labels: dictionary of labels to apply on pod
+        :param tolerations: dictionary tolerations for nodes
+        :param annotations: dictionary of annotations to apply on pod
         :param source: mount point of shared storage
         """
         self._task_id = task_id
@@ -97,10 +101,10 @@ class NodePodOperator(KubernetesPodOperator):
             startup_timeout_seconds=startup_timeout,
             is_delete_operator_pod=True,
             pod_template_file=self.minimal_pod_template,
-            node_selector=node_selector_labels,
             node_selectors=node_selector_labels,
             labels=labels,
-            tolerations=tolerations
+            tolerations=self.create_tolerations(tolerations),
+            annotations=annotations
         )
 
     @staticmethod
@@ -174,3 +178,23 @@ spec:
             if not volume_disabled
             else k8s.V1PodSecurityContext()
         )
+
+    @staticmethod
+    def create_tolerations(
+        tolerations: Optional[List[Dict[str, str]]] = None
+    ) -> List[k8s.V1Toleration]:
+        """
+        Creates k8s tolerations
+        :param tolerations:
+        :return:
+        """
+        if not tolerations:
+            return []
+        return [
+            k8s.V1Toleration(
+                effect=toleration.get("effect"),
+                key=toleration.get("key"),
+                operator=toleration.get("operator"),
+                value=toleration.get("value"))
+            for toleration in tolerations
+        ]